requirements/base.txt
asgiref==3.5.2
autopep8==1.6.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bootstrap4==0.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.1.5
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
exponent-server-sdk==2.0.0
filelock==3.7.1
idna==3.4
inflection==0.5.1
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.16.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.5
psycopg2-binary==2.9.3
ptyprocess==0.6.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.6.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
toml==0.10.2
traitlets==4.3.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.14.1
wcwidth==0.1.9
